requirements.txt
python-telegram-bot==21.7
httpx==0.24.1
